Replaced manual newline technique for description lists.
In two descriptions lists of [cpp.predefined], a newline is added manually
after the description text. This requires putting the defined term in the
definition, i.e. outside the optional argument of \item.

This commit replaces this technique to insert a newline by using the
enumitem package and automatic line breaks.